changing name of user controller to UserVerificationController.py
